Added dropdown check box
